Added Users editing functionality to Profile Page.

+++ server/routes/users/deleteUser.js
@@ -2,9 +2,14 @@ const StudentModel = require('../../models/student.model');
 
 module.exports = async (req, res) => {
   const { id } = req.params;
-  const student = await StudentModel.findById(id);
 
-  await student.deleteOne();
-  res.status(204).json(student);
+  try {
+    const student = await StudentModel.findById(id);
+    if (!student) return;
+
+    await student.deleteOne();
+    res.status(204).json({});
+  } catch (e) {
+    console.log('Cannot delete user ', e);
+  }
 };
-
